rework of classification & group classification constructor & static convience methods; main use methods are now on the interface (API users will need to pass in a constructor method).
this eliminates the duplication on every single impl class of these highly useful
methods.

there is still a lot of repeat on the *interfaces*; basically to implement
cogent type-casting. could not come up with a way to avoid that;
but at least we can easily add an impl type now without having to do a 100 lines
of boilerplate